pyproject.toml
[tool.poetry]
name = "ok1newsbot"
version = "0.1.0"
description = ""
authors = ["Vadym Lishchynskyi <ya.lishinskiy2017@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
telethon = "^1.27.0"
python-dotenv = "^1.0.0"
python-telegram-bot = "^20.1"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
